+++ copa/juegos/index.js
@@ -9,15 +9,19 @@
 import * as linea from './linea.js';
 import * as numero from './numero.js';
 import * as conexiones from './conexiones.js';
-import * as solitario from './solitario.js';
-import * as dudo from './dudo.js';
+import * as reinas from './reinas.js';
+import * as letras from './letras.js';
+import * as zip from './zip.js';
+import * as tango from './tango.js';
 import * as anio from './anio.js';
 import * as final from './final.js';
 import * as uiLinea from './ui-linea.js';
 import * as uiNumero from './ui-numero.js';
 import * as uiConexiones from './ui-conexiones.js';
-import * as uiSolitario from './ui-solitario.js';
-import * as uiDudo from './ui-dudo.js';
+import * as uiReinas from './ui-reinas.js';
+import * as uiLetras from './ui-letras.js';
+import * as uiZip from './ui-zip.js';
+import * as uiTango from './ui-tango.js';
 import * as uiAnio from './ui-anio.js';
 import * as uiFinal from './ui-final.js';
 
@@ -27,8 +31,10 @@ export const JUEGOS = {
   linea: juego(linea, uiLinea),
   numero: juego(numero, uiNumero),
   conexiones: juego(conexiones, uiConexiones),
-  solitario: juego(solitario, uiSolitario),
-  dudo: juego(dudo, uiDudo),
+  reinas: juego(reinas, uiReinas),
+  letras: juego(letras, uiLetras),
+  zip: juego(zip, uiZip),
+  tango: juego(tango, uiTango),
   anio: juego(anio, uiAnio),
   final: juego(final, uiFinal),
 };